Add requirements.txt
This commit is not suitable for upstreaming because it is SeaGL-
specific. The upstream bug to fix in a general way (which will also
work for dev environments, etc.) is pretalx/pretalx#1747.

requirements.txt was generated with:

pip3 install --upgrade-strategy=eager -Ue ".[postgres, redis]" gunicorn
pip3 freeze | grep -v pretalx.git > requirements.txt

I also happen to have pip-tools in the venv and pip-sync does not
remove itself. Who cares.